notes and notes setting seaperated
