+++ src/repository/storage/postgres/orm/sequelize/note.ts
@@ -4,8 +4,6 @@ import type Orm from '@repository/storage/postgres/orm/sequelize/index.js';
 import type { Note, NoteInternalId, NotePublicId } from '@domain/entities/note.js';
 import type { NoteCreationAttributes } from '@domain/entities/note.js';
 import { NotesSettingsModel } from '@repository/storage/postgres/orm/sequelize/notesSettings.js';
-import type NotesSettings from '@domain/entities/notesSettings.js';
-import type { NotesSettingsCreationAttributes } from '@domain/entities/notesSettings.js';
 import { UserModel } from '@repository/storage/postgres/orm/sequelize/user.js';
 
 /* eslint-disable @typescript-eslint/naming-convention */
@@ -71,16 +69,11 @@ export default class NoteSequelizeStorage {
   private readonly tableName = 'notes';
 
   /**
-   * Settings table name
-   */
-  private readonly settingsTableName = 'notes_settings';
-
-  /**
    * Constructor for note storage
    *
    * @param ormInstance - ORM instance
    */
-  constructor({ connection }: Orm) {
+  constructor({ connection }: Orm, settingModal: typeof NotesSettingsModel) {
     this.database = connection;
 
     /**
@@ -116,38 +109,13 @@ export default class NoteSequelizeStorage {
     /**
      * Initiate note settings model
      */
-    this.settingsModel = NotesSettingsModel.init({
-      id: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true,
-      },
-      note_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        references: {
-          model: NoteModel,
-          key: 'id',
-        },
-      },
-      custom_hostname: {
-        type: DataTypes.STRING,
-        allowNull: true,
-      },
-      enabled: {
-        type: DataTypes.BOOLEAN,
-        allowNull: false,
-        defaultValue: true,
-      },
-    }, {
-      tableName: this.settingsTableName,
-      sequelize: this.database,
-      timestamps: false,
-    });
+    this.settingsModel = settingModal;
 
     /** NoteModel and NotesSettingsModel are connected as ONE-TO-ONE */
-    this.model.hasOne(this.settingsModel, { foreignKey: 'note_id',
-      as: this.settingsModel.tableName });
+    this.model.hasOne(this.settingsModel, {
+      foreignKey: 'note_id',
+      as: this.settingsModel.tableName
+    });
   }
 
   /**
@@ -213,34 +181,6 @@ export default class NoteSequelizeStorage {
   }
 
   /**
-   * Gets note settings by id
-   *
-   * @param id - note id
-   * @returns { Promise<NotesSettings | null> } found note
-   */
-  public async getNoteSettingsById(id: NotesSettings['id']): Promise<NotesSettings | null> {
-    const noteSettings = await this.settingsModel.findOne({
-      where: {
-        id,
-      },
-    });
-
-    /**
-     * If note settings not found, return null
-     */
-    if (!noteSettings) {
-      return null;
-    }
-
-    return {
-      id: noteSettings.id,
-      noteId: noteSettings.note_id,
-      enabled: noteSettings.enabled,
-      customHostname: noteSettings.custom_hostname,
-    };
-  }
-
-  /**
    * Gets note by id
    *
    * @param hostname - custom hostname
@@ -279,120 +219,4 @@ export default class NoteSequelizeStorage {
 
     return note;
   };
-
-  /**
-   * Get note settings
-   *
-   * @param noteId - note id
-   * @returns { Promise<NotesSettings | null> } - note settings
-   */
-  public async getNoteSettingsByNoteId(noteId: NotesSettings['noteId']): Promise<NotesSettings> {
-    const settings = await this.settingsModel.findOne({
-      where: {
-        note_id: noteId,
-      },
-    });
-
-    if (!settings) {
-      /**
-       * TODO: improve exceptions
-       */
-      throw new Error('Note settings not found');
-    }
-
-    return {
-      id: settings.id,
-      noteId: settings.note_id,
-      customHostname: settings.custom_hostname,
-      enabled: settings.enabled,
-    };
-  }
-
-  /**
-   * Get note settings
-   *
-   * @param id - note internal id
-   * @returns { Promise<NotesSettings | null> } - note settings
-   *
-   * @deprecated
-   * @todo resolve note setting by internal id
-   */
-  public async getNoteSettingsByPublicId(id: NotePublicId): Promise<NotesSettings> {
-    const settings = await this.settingsModel.findOne({
-      where: {
-        id: id,
-      },
-    });
-
-    if (!settings) {
-      /**
-       * TODO: improve exceptions
-       */
-      throw new Error('Note settings not found');
-    }
-
-    return {
-      id: settings.id,
-      noteId: settings.note_id,
-      customHostname: settings.custom_hostname,
-      enabled: settings.enabled,
-    };
-  }
-
-  /**
-   * Insert note settings
-   *
-   * @param options - note settings options
-   * @returns { Promise<NotesSettings> } - inserted note settings
-   */
-  public async insertNoteSettings({
-    noteId,
-    customHostname,
-    enabled,
-  }: NotesSettingsCreationAttributes
-  ): Promise<NotesSettings> {
-    const settings = await this.settingsModel.create({
-      note_id: noteId,
-      custom_hostname: customHostname,
-      enabled: enabled,
-    });
-
-    return {
-      id: settings.id,
-      noteId: settings.note_id,
-      customHostname: settings.custom_hostname,
-      enabled: settings.enabled,
-    };
-  }
-
-  /**
-   * Update note settings
-   *
-   * @param data - note settings new data
-   * @param id - note settings id
-   */
-  public async patchNoteSettings(data: Partial<NotesSettings>, id: NotesSettings['id']): Promise<NotesSettings | null> {
-    const settingsToUpdate = await this.settingsModel.findByPk(id);
-
-    /**
-     * Check if note settings exists in database
-     */
-    if (!settingsToUpdate) {
-      return null;
-    }
-
-    const values = {
-      enabled: data.enabled,
-      custom_hostname: data.customHostname,
-    };
-
-    const updatedSettings = await settingsToUpdate.update(values);
-
-    return {
-      id: updatedSettings.id,
-      noteId: updatedSettings.note_id,
-      customHostname: updatedSettings.custom_hostname,
-      enabled: updatedSettings.enabled,
-    };
-  }
 }
